main layout and new routes
