+++ src/config/index.ts
@@ -1,0 +1,18 @@
+import type { MenuItem } from '@/components/Menu/OMenuItem.vue'
+
+export const menu: MenuItem[] = [
+  {
+    name: 'Administrador',
+    items: [
+      {
+        icon: 'star',
+        name: 'Talento',
+        items: [
+          {
+            name: 'Reclutamiento'
+          }
+        ]
+      }
+    ]
+  }
+]
